Move db to the internal module
