+++ bellhop/src/internal/mod.rs
@@ -1,2 +1,3 @@
 pub mod auth;
+pub mod db;
 pub mod hooks;
