+++ bellhop/src/views/endpoints.rs
@@ -1,4 +1,4 @@
-use crate::db::Db;
+use crate::internal::db::Db;
 use crate::errors::*;
 use crate::internal::hooks::Hooks;
 use crate::sheriff;
